Added Linear Regression take 2 (#12)
Added linear regression. Added Coefficient of Determination, so we can project how the data will preform in the future, using a measure of confidence, see the sample.html. Added projections using confidence, updated md to show usage, check out the sample.html for demo. Added the slope, intercept and r^2 to the hmtl for completeness. Added notation on returns.